requirements.txt
mistralai==0.0.7
python-dotenv==1.0.0
